Rework docker stuff.
