+++ rosetta-server-polkadot/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "rosetta-server-polkadot"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.66"
+clap = { version = "4.0.27", features = ["derive"] }
+femme = "2.2.1"
+rosetta-server-substrate = { path = "../rosetta-server-substrate" }
+tide = "0.16.0"
+tokio = { version = "1.21.2", features = ["full"] }
