+++ rosetta-server-substrate/Cargo.toml
@@ -8,7 +8,6 @@ anyhow = "1.0.65"
 array-bytes = "4.1.0"
 blake2-rfc = "0.2.18"
 bs58 = "0.4.0"
-femme = "2.2.1"
 hex = "0.4.3"
 parity-scale-codec = "3.2.1"
 rosetta-types = { path = "../rosetta-types" }
@@ -17,7 +16,5 @@ serde_json = "1.0.87"
 sp-keyring = "6.0.0"
 ss58-registry = "1.31.0"
 subxt = "0.24.0"
-subxt-codegen = "0.25.0"
 syn = "1.0.80"
 tide = "0.16.0"
-tokio = { version = "1.21.2", features = ["full"] }
